Fix job

+++ app/Jobs/TranscodeFromS3Job.php
@@ -43,7 +43,10 @@ class TranscodeFromS3Job implements ShouldQueue
         TranscodeVideoJob::dispatchNow($this->partVideo);
 
         printf("Delete file on s3.\n");
-        Storage::disk('s3')->delete($this->partVideo->s3_path . '/input.tmp');
+        $this->partVideo->refresh();
+        if ($this->partVideo->transcode_status == TranscodeStatus::COMPLETED) {
+            Storage::disk('s3')->delete($this->partVideo->s3_path . '/input.tmp');
+        }
     }
 
     public function failed(Exception $ex)
